+++ src/bakery_manager_mobile/lib/assets/constants.dart
@@ -1,0 +1,11 @@
+const String homePageRoute = '/';
+const String recipePageRoute = '/recipes';
+const String ingredientPageRoute = '/ingredients';
+const String productPageRoute = '/products';
+const String taskPageRoute = '/tasks';
+const String salesPageRoute = '/sales';
+const String loginPageRoute = '/login';
+const String registrationPageRoute = '/login/registration';
+const String addRecipePageRoute = '/recipes/addRecipe';
+const String recipeDetailsPageRoute = '/recipes/details';
+const String myAccountRoute = '/myaccount';