+++ CHANGELOG.md
@@ -7,7 +7,33 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-- Nothing for now
+- nothing
+
+## [0.1.0] - 2024-09-19
+
+### Added
+
+- Added Page Routing Between Pages. 
+- Added Const File to Hold Constants Used in Multiple Files.
+- Added Intitial Project Src 
+- Added Static Home Page 
+- Created a login page.
+- Added custom fonts.
+- Added api service to handle api calls.
+- Added api calls for login, register, get recipes, and add recipe.
+- Added navigator observer to handle to reload when navigating to a back to a page.
+- Added a custom launcher icon.
+- Added recipe details page for displaying recipe details.
+- Added Account Page for managing user account.
+- Added Mock ingredients page
+- Added Mock task page
+- Added API Routes for recipe ingredients
+
+### Changed
+
+- Moved Most Code Outside main.dart Into Individual Files Based off Page.
+- Updated Changelog with resources for ease of access
+- Moved add recipe page to its own file.
 
 ## [0.0.2] - 2024-08-28
 
@@ -21,5 +47,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Initial Commit
 
+[Unreleased]: https://github.com/Swolford0408/MobileDev/compare/v0.1.0...HEAD
+[0.1.0]: https://github.com/Swolford0408/MobileDev/compare/v0.0.2...0.1.0
 [0.0.2]: https://github.com/Swolford0408/MobileDev/compare/v0.0.1...v0.0.2
 [0.0.1]: https://github.com/Swolford0408/MobileDev/releases/tag/v0.0.1